let stream info requests work with no vehicle connected, change SessionMetadata time properties to use quint64 instead

+++ src/Comms/LogReplayLinkController.h
@@ -156,8 +156,8 @@ private:
 
     struct SessionMetadata {
         QString flightId;
-        uint32_t dcDurationSecs;
-        uint32_t tlogDurationSecs;
+        quint64 dcDurationSecs;
+        quint64 tlogDurationSecs;
         bool hasVideo;
         bool hasTlog;
         QString tlogFilePath;
